+++ Sygic.Corona.Domain/Alert.cs
@@ -6,11 +6,11 @@ namespace Sygic.Corona.Domain
     {
         public Guid Id { get; private set; }
         public string DeviceId { get; private set; }
-        public uint ProfileId { get; private set; }
+        public long ProfileId { get; private set; }
         public DateTime CreatedOn { get; private set; }
         public string Content { get; private set; }
 
-        public Alert(string deviceId, uint profileId, DateTime createdOn, string content)
+        public Alert(string deviceId, long profileId, DateTime createdOn, string content)
         {
             DeviceId = deviceId;
             ProfileId = profileId;
